+++ module/behaviour/strategy/SoccerStrategy/src/SoccerStrategy.hpp
@@ -43,30 +43,30 @@ namespace module::behaviour::strategy {
             Config() = default;
             NUClear::clock::duration ball_last_seen_max_time{};
             NUClear::clock::duration goal_last_seen_max_time{};
-            float ball_search_walk_start_speed = 0.0f;
-            float ball_search_walk_stop_speed  = 0.0f;
-            float ball_search_walk_slow_time   = 0.0f;
+            double ball_search_walk_start_speed = 0.0f;
+            double ball_search_walk_stop_speed  = 0.0f;
+            double ball_search_walk_slow_time   = 0.0f;
             Eigen::Vector2d start_position_offensive{Eigen::Vector2d::Zero()};
             Eigen::Vector2d start_position_defensive{Eigen::Vector2d::Zero()};
-            bool is_goalie                         = false;
-            float goalie_max_ball_distance         = 0.0f;
-            float goalie_command_timeout           = 0.0f;
-            float goalie_rotation_speed_factor     = 0.0f;
-            float goalie_max_rotation_speed        = 0.0f;
-            float goalie_translation_speed_factor  = 0.0f;
-            float goalie_max_translation_speed     = 0.0f;
-            float goalie_side_walk_angle_threshold = 0.0f;
+            bool is_goalie                          = false;
+            double goalie_max_ball_distance         = 0.0f;
+            double goalie_command_timeout           = 0.0f;
+            double goalie_rotation_speed_factor     = 0.0f;
+            double goalie_max_rotation_speed        = 0.0f;
+            double goalie_translation_speed_factor  = 0.0f;
+            double goalie_max_translation_speed     = 0.0f;
+            double goalie_side_walk_angle_threshold = 0.0f;
             NUClear::clock::duration localisation_interval{};
             NUClear::clock::duration localisation_duration{};
-            bool force_playing               = false;
-            bool force_penalty_shootout      = false;
-            int walk_to_ready_time           = 0;
-            float kicking_distance_threshold = 0.0f;
-            float kicking_angle_threshold    = 0.0f;
+            bool force_playing                = false;
+            bool force_penalty_shootout       = false;
+            int walk_to_ready_time            = 0;
+            double kicking_distance_threshold = 0.0f;
+            double kicking_angle_threshold    = 0.0f;
         } cfg;
 
         /// @brief Flag to determine whether the ball is on the left or right side of the robot
-        float ball_lost_clockwise = true;
+        double ball_lost_clockwise = true;
 
         /// @brief Bool to indicate  if the robot is currently getting up
         bool is_getting_up = false;
